query_converter/requirements.txt
asgiref==3.7.2
black==24.3.0
cfgv==3.4.0
click==8.1.7
coverage==7.4.4
distlib==0.3.8
Django==5.0.3
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.15.0
exceptiongroup==1.2.0
Faker==24.3.0
filelock==3.13.1
flake8==7.0.0
identify==2.5.35
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy==1.9.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.2
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.20240311
typing_extensions==4.10.0
virtualenv==20.25.1
